+++ packages/components/src/assets/IconProps.ts
@@ -1,3 +1,0 @@
-export type IconProps = {
-    size?: number
-}