+++ packages/core/src/components/ui/Button/ButtonSignIn/ButtonSignIn.tsx
@@ -1,16 +1,19 @@
-import { LinkButton } from '@faststore/ui'
+import { Icon, LinkButton } from '@faststore/ui'
 
 import styles from 'src/components/ui/Button/button.module.scss'
-import Icon from 'src/components/ui/Icon'
 import { useSession } from 'src/sdk/session'
+import { useCart } from '../../../../sdk/cart/index'
 
 const ButtonSignIn = () => {
+  const { id } = useCart()
   const { person } = useSession()
 
   return (
     <LinkButton
       data-fs-button-signin-link
-      href={person?.id ? '/account' : '/login'}
+      href={
+        person?.id ? `/account?orderFormId=${id}` : `/login?orderFormId=${id}`
+      }
       className={`${styles.fsButton} text__title-mini`}
       variant="tertiary"
       icon={<Icon name="User" width={18} height={18} weight="bold" />}