+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.98-alpha.0](https://github.com/vtex/faststore/compare/v2.0.97-alpha.0...v2.0.98-alpha.0) (2023-04-03)
+
+### Chores
+
+- updates api version in core package ([#1702](https://github.com/vtex/faststore/issues/1702)) ([d2bbfae](https://github.com/vtex/faststore/commit/d2bbfae551417cd6ac525fa2fe8768956254b6ec))
+
+## [2.0.97-alpha.0](https://github.com/vtex/faststore/compare/v2.0.96-alpha.0...v2.0.97-alpha.0) (2023-04-03)
+
+**Note:** Version bump only for package faststore
+
+## [2.0.96-alpha.0](https://github.com/vtex/faststore/compare/v2.0.95-alpha.0...v2.0.96-alpha.0) (2023-04-03)
+
+### Bug Fixes
+
+- Fix useDefineVariants ([#1700](https://github.com/vtex/faststore/issues/1700)) ([463b65b](https://github.com/vtex/faststore/commit/463b65b1e9fb81ca041f7f7f24280a57e3210b9b))
+
+## [2.0.95-alpha.0](https://github.com/vtex/faststore/compare/v2.0.94-alpha.0...v2.0.95-alpha.0) (2023-04-01)
+
+### Features
+
+- Adds `RegionBar` component ([#1693](https://github.com/vtex/faststore/issues/1693)) ([8390d82](https://github.com/vtex/faststore/commit/8390d829d4c1cab0f56a38b59bb2e1f1ab670fb1))
+
 ## [2.0.94-alpha.0](https://github.com/vtex/faststore/compare/v2.0.93-alpha.0...v2.0.94-alpha.0) (2023-03-31)
 
 ### Chores